Update install.md
I would consider swapping order of the sections on Comets and Planets.  I may be mistaken, but I think far more people enter Urbit via the planet flow than with a comet. 

Also, should we add a "Live Install Support" section with links to the various Telegram and Discord groups?