docs: CLI Install overwrites old binary
The recent change to the Command Line install instructions introduced
the usage of tar's -k flag, which prevents overwriting any existing
file with the same name. As such, some users were having trouble
upgrading vere in place, because copy and pasting the command seemingly
did nothing.

This restores the previous behavior: copy-and-paste the command to
download the latest binary, even if one is already present.